requirements.txt
tqdm>=4.65.0
scikit-learn>=1.3.0
matplotlib>=3.7.4
networkx>=3.1
scikit-image~=0.19.3
numpy>=1.24.3
torchvision>=0.13.0
trimesh>=4.0.8
pillow~=10.4.0